+++ packages/site/src/content/InputCurrency/index.tsx
@@ -1,6 +1,7 @@
 import Content from "@atlantis/docs/components/InputCurrency/InputCurrency.stories.mdx";
 import MobileProps from "./InputCurrency.props-mobile.json";
 import { ContentExport } from "../../types/content";
+import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
 export default {
   content: () => <Content />,
@@ -12,7 +13,9 @@ export default {
   links: [
     {
       label: "Storybook",
-      url: "http://localhost:6006/?path=/docs/components-utilities-InputCurrency-web--docs",
+      url: getStorybookUrl(
+        `?path=/docs/components-forms-and-inputs-inputcurrency--docs`,
+      ),
     },
   ],
 } as const satisfies ContentExport;